docs(components/molecule/radioButtonGroup): refactor
